requirements.txt
Flask==0.9
Flask-PyMongo==0.2.1
simplejson==2.6.2
validictory==0.8.3
